Added piece of code to enable trophies feat in read.me of Aayush
Aayush, This is Small piece of addition i thought, if you ever think of giving this a try..by seeing trophies in other's read.me file. So being a friendi thought to give it away. You merge this in ..and see how does it look , u may comment this feat later as u suit.